[refacto] decomposition the Profile component #92
