+++ packages/client/src/pages/Profile/Settings.tsx
@@ -1,0 +1,5 @@
+import { Button } from 'antd'
+
+export const Settings = () => {
+  return <Button>Сменить пароль</Button>
+}
