+++ packages/client/src/pages/Profile/data.ts
@@ -1,0 +1,27 @@
+export const Fields = [
+  {
+    placeholder: 'Имя',
+    name: 'first_name',
+    editable: true,
+  },
+  {
+    placeholder: 'Фамилия',
+    name: 'second_name',
+    editable: true,
+  },
+  {
+    placeholder: 'Email',
+    name: 'email',
+    editable: true,
+  },
+  {
+    placeholder: 'Телефон',
+    name: 'phone',
+    editable: true,
+  },
+  {
+    placeholder: 'Логин',
+    name: 'login',
+    editable: true,
+  },
+]
